r18301: I discovered how to load the warnings from a build farm build into emacs compile mode (hint, paste to a file, and compile as "cat filename").
This allowed me to fix nearly all the warnings for a IA_64 SuSE build
very quickly.

+++ source/lib/appweb/mpr/var.c
@@ -1672,7 +1672,7 @@ void mprVarToString(char** out, int size, char *fmt, MprVar *obj)
 #if BLD_GOAHEAD_WEBSERVER
 			mprAllocSprintf(out, size, "%d", (int) obj->integer64);
 #else
-			mprAllocSprintf(out, size, "%lld", obj->integer64);
+			mprAllocSprintf(out, size, "%lld", (long long)obj->integer64);
 #endif
 		} else {
 			mprAllocSprintf(out, size, fmt, obj->integer64);
